requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
arrow==1.2.2
    # via django-q
asgiref==3.5.0
    # via django
blessed==1.19.1
    # via django-q
boto3==1.24.17
    # via -r requirements.in
botocore==1.27.17
    # via
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via whitenoise
certifi==2022.6.15
    # via requests
charset-normalizer==2.0.12
    # via requests
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
django==4.0.4
    # via
    #   -r requirements.in
    #   django-address
    #   django-drf-filepond
    #   django-extensions
    #   django-import-export
    #   django-model-utils
    #   django-modeltranslation
    #   django-phonenumber-field
    #   django-picklefield
    #   django-q
    #   django-sesame
    #   django-storages
    #   djangorestframework
django-address==0.2.6
    # via -r requirements.in
django-constance[database]==2.9.0
    # via -r requirements.in
django-countries==7.3.2
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via -r requirements.in
django-drf-filepond==0.4.1
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-import-export==2.8.0
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-modeltranslation==0.17.5
    # via -r requirements.in
django-phonenumber-field[phonenumberslite]==6.1.0
    # via -r requirements.in
django-picklefield==3.1
    # via
    #   django-constance
    #   django-q
django-q==1.3.9
    # via
    #   -r requirements.in
    #   django-q-email
django-q-email==5.0.1
    # via -r requirements.in
django-sesame[ua]==2.4
    # via -r requirements.in
django-storages==1.12.3
    # via
    #   -r requirements.in
    #   django-drf-filepond
djangorestframework==3.13.1
    # via django-drf-filepond
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markuppy==1.14
    # via tablib
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
phonenumberslite==8.12.46
    # via django-phonenumber-field
pillow==9.1.1
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
pymemcache==3.5.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
pytz==2022.1
    # via djangorestframework
pyyaml==6.0
    # via tablib
redis==3.5.3
    # via django-q
requests==2.28.0
    # via django-drf-filepond
s3transfer==0.6.0
    # via boto3
shortuuid==1.0.9
    # via django-drf-filepond
six==1.16.0
    # via
    #   blessed
    #   django-drf-filepond
    #   django-modeltranslation
    #   pymemcache
    #   python-dateutil
sqlparse==0.4.2
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
typing-extensions==4.1.1
    # via django-countries
ua-parser==0.10.0
    # via django-sesame
urllib3==1.26.9
    # via
    #   botocore
    #   requests
wcwidth==0.2.5
    # via blessed
whitenoise[brotli]==6.0.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
